@joshcooper 2021-05-13
AFAIK we don't ship ruby 2.6 in any of our vanagon projects?

I use procmon to capture events for "Process Name = ruby.exe" and then under 'Tools' I think there's an option to view File Summary, and lists total, read, write, etc events.