@GabrielNagy 2021-05-13
> AFAIK we don't ship ruby 2.6 in any of our vanagon projects?

Ah sorry, was thinking about puppet6/ruby2.5