@joshcooper 2021-05-13
ah yeah, that's possible. I was thinking after releasing 7.7.0, we could look into backporting some of these perf changes to 6, especially https://github.com/puppetlabs/puppet-runtime/commit/c56420a650cfd5c3188fdbbdc6781ce886a193b0